requirements.txt
# Web Framework
fastapi==0.115.2
uvicorn[standard]==0.32.0

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
# Для Postgres
psycopg2-binary==2.9.10

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
python-multipart==0.0.9

# Testing
pytest==8.3.3
httpx==0.27.2
pre-commit

sqlalchemy